Linting the Import and Export features
Renamed the package and files so as to be more descriptive but still
avoid potential namespacing issues on case-insensitive OSes.
Have added comments to exported functions and renamed the package and
files to be consistent with importPkg.
Linted the Request package and fixing comments in other files
Linted the Utils package